+++ src/main.rs
@@ -4,14 +4,9 @@ extern crate rand;
 pub(crate) mod vlang_src_load;
 mod frontend;
 
+#[cfg(test)]
+mod test;
+
 fn main() {
     frontend::run();
-}
-
-#[cfg(test)]
-mod main_test {
-    #[test]
-    fn is_run() {
-        assert_eq!(1 + 1, 2);
-    }
 }
