+++ src/test/main_test.rs
@@ -1,0 +1,4 @@
+#[test]
+fn is_run() {
+    assert_eq!(1 + 1, 2);
+}
